Week 3. Day 2. Auto Layout and Cells Exercise

+++ Week 3/Day 2/AutoLayoutCell/AutoLayoutCell/BusinessCell.m
@@ -43,7 +43,7 @@
     [self.thumbImageView setImageWithURL:[NSURL URLWithString:self.business.imageUrl]];
     self.nameLabel.text = self.business.name;
     [self.ratingImageView setImageWithURL:[NSURL URLWithString:self.business.ratingImageUrl]];
-    self.ratingLabel.text = [NSString stringWithFormat:@"%ld Reviews", self.business.numReviews];
+    self.ratingLabel.text = [NSString stringWithFormat:@"%ld Reviews", (long)self.business.numReviews];
     self.addressLabel.text = self.business.address;
     self.distanceLabel.text = [NSString stringWithFormat:@"%.2f mi", self.business.distance];
     self.snippetLabel.text = self.business.snippet;
